+++ test/integration/api/reports/unchecked_test.rb
@@ -1,0 +1,18 @@
+# frozen_string_literal: true
+
+require 'test_helper'
+
+class API::Reports::UncheckedTest < ActionDispatch::IntegrationTest
+  fixtures :products
+
+  test 'GET /api/reports/unchecked/counts.txt' do
+    get counts_api_reports_unchecked_index_path(format: :text)
+    assert_response :unauthorized
+
+    token = create_token('komagata', 'testtest')
+    get counts_api_reports_unchecked_index_path(format: :text),
+        headers: { 'Authorization' => "Bearer #{token}" }
+    assert_response :ok
+    assert_match '63件', response.body
+  end
+end
