+++ test/integration/api/reports_test.rb
@@ -35,15 +35,4 @@ class API::ReportsTest < ActionDispatch::IntegrationTest
         headers: { 'Authorization' => "Bearer #{token}" }
     assert_response :ok
   end
-
-  test 'GET /api/reports/unchecked/counts.txt' do
-    get counts_api_reports_unchecked_index_path(format: :text)
-    assert_response :unauthorized
-
-    token = create_token('komagata', 'testtest')
-    get counts_api_reports_unchecked_index_path(format: :text),
-        headers: { 'Authorization' => "Bearer #{token}" }
-    assert_response :ok
-    assert_match '63件', response.body
-  end
 end
